+++ src/test/java/lotto/model/MoneyTest.java
@@ -11,7 +11,7 @@ public class MoneyTest {
 	@Test
 	@DisplayName("금액은 음수가 될 수 없음")
 	void validateNegativeNumber() {
-		int negativeNumber = -1;
+		long negativeNumber = -1L;
 		assertThatIllegalArgumentException().isThrownBy(() -> {
 			Money money	= new Money(negativeNumber);
 		});
@@ -20,12 +20,12 @@ public class MoneyTest {
 	@Test
 	@DisplayName("금액은 음이 아닌 정수")
 	void validateNonNegativeNumber() {
-		int zero = 0;
+		long zero = 0L;
 		assertThatNoException().isThrownBy(() -> {
 			Money money	= new Money(zero);
 		});
 
-		int positiveNumber = 1;
+		long positiveNumber = 1L;
 		assertThatNoException().isThrownBy(() -> {
 			Money money	= new Money(positiveNumber);
 		});
